Move spin functions to executor module
